pyproject.toml
[tool.poetry]
name = "paseto"
version = "1.0.0"
description = "Platform-Agnostic Security Tokens for Python (PASETO)"
authors = ["Ryan Littlefield <ryan@ryanlittlefield.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pendulum = "^2.1.2"
pyseto = "^0.6.1"
pysodium = "^0.7.10"
pycryptodomex = "^3.10.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
